Save a Giotto object to an AnnData-Zarr store

Save a Giotto object to an AnnData-Zarr store

Usage

giotto_to_anndata_zarr(giotto_obj, out_path, X_slot = "raw_exprs")

Arguments

giotto_obj

The object to save.

out_path

A path to the output Zarr store.

X_slot

The name of the slot in the Giotto object to use for adata.X

Value

TRUE if the conversion succeeds.

Examples

obj <- get_giotto_obj()
#> Consider to install these (optional) packages to run all possible Giotto commands for spatial analyses:  scran MAST smfishHmrf trendsceek SPARK multinet RTriangle FactoMiner
#>  Giotto does not automatically install all these packages as they are not absolutely required and this reduces the number of dependencies
#>  no external python path or giotto environment was specified, will check if a default python path is available 
#> 
#>  A default python path was found:  /usr/bin/python3  and will be used
#> 
#>  If this is not the correct python path, either
#> 
#>  1. use installGiottoEnvironment() to install a local miniconda python environment along with required modules 
#> 
#>  2. provide an existing python path to python_path to use your own python path which has all modules installed 
#> Warning: module: pandas was not found with python path: /usr/bin/python3
#> Warning: module: igraph was not found with python path: /usr/bin/python3
#> Warning: module: leidenalg was not found with python path: /usr/bin/python3
#> Warning: module: community was not found with python path: /usr/bin/python3
#> Warning: module: networkx was not found with python path: /usr/bin/python3
#> Warning: module: sklearn was not found with python path: /usr/bin/python3
giotto_to_anndata_zarr(obj, "data/giotto.zarr")
#> [1] TRUE